Add comments.

+++ GKA1/src/gka1/GkaGraph.java
@@ -228,8 +228,8 @@ public class GkaGraph extends MultiGraph {
 		double min = Double.POSITIVE_INFINITY;
 		for (Edge edge : sourceNode.getEachEdge()) {
  			if (sourceNode.hasEdgeToward(targetNode) && sourceNode.getEdgeToward(targetNode).equals(edge)) {				
- 				if (Integer.valueOf(edge.getAttribute("weight").toString()) < min) {
- 					min = Integer.valueOf(edge.getAttribute("weight").toString());
+ 				if (Double.valueOf(edge.getAttribute("weight").toString()) < min) {
+ 					min = Double.valueOf(edge.getAttribute("weight").toString());
  				}
  			}
  		}
